// src/app/core/api/model/authResponseDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AuthResponseDTO { 
    accessToken?: string;
    refreshToken?: string;
    userId?: string;
    email?: string;
}