// src/app/core/api/model/appUserDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AppUserDTO { 
    id?: string;
    email: string;
    password: string;
    createdAt?: string;
}